+++ R/keyring.R
@@ -10,20 +10,37 @@ TOKEN <- "fb_adlib_token"
 APP_ID <- "fb_adlib_app_id"
 APP_SECRET <- "fb_adlib_app_secret"
 
+# @importFrom rstudioapi versionInfo
+is_rstudio_server <- function() {
+  rstudioapi::versionInfo()$mode == "server"
+}
+
+get_pass <- function(msg) {
+  if (is_rstudio_server()) {
+    out <- readline(paste0(msg, ": "))
+  } else {
+    out <- getPass::getPass(msg, noblank = T, forcemask = F)
+  }
+  return(out)
+}
 
 is_set_up <- function() {
   return(secret_exists(TOKEN) & secret_exists(APP_ID) & secret_exists(APP_SECRET))
 }
 
+keyring_backend_is_env <- function() {
+  suppressWarnings(keyring::default_backend()$name == "env")
+}
+
 secret_set <- function(secret = c(TOKEN, APP_ID, APP_SECRET), value) {
   secret <- match.arg(secret)
-  keyring::key_set_with_value(service = SERVICE, username = secret, password = value)
+  suppressWarnings(keyring::key_set_with_value(service = SERVICE, username = secret, password = value))
 }
 
 secret_get <- function(secret = c(TOKEN, APP_ID, APP_SECRET)) {
   secret <- match.arg(secret)
   if (secret_exists(secret)) {
-    return(keyring::key_get(service = SERVICE, username = secret))
+    return(suppressWarnings(keyring::key_get(service = SERVICE, username = secret)))
   }
   stop(glue::glue("Secret {secret} does not exist in keychain."))
 }
@@ -44,7 +61,7 @@ adlib_clear_setup <- function() {
 
 secret_delete <- function(secret = c(TOKEN, APP_ID, APP_SECRET)) {
   secret <- match.arg(secret)
-  keyring::key_delete(service = SERVICE, username = secret)
+  suppressWarnings(keyring::key_delete(service = SERVICE, username = secret))
 }
 
 token_set <- function(token) {
@@ -85,7 +102,12 @@ token_delete <- function() {
 
 secret_exists <- function(secret = c(TOKEN, APP_ID, APP_SECRET)) {
   secret <- match.arg(secret)
-  secret %in% keyring::key_list(service = SERVICE)[["username"]]
+  suppressWarnings(secret %in% keyring::key_list(service = SERVICE)[["username"]])
+}
+
+
+yesno <- function(msg) {
+  menu(c("yes", "no"), title = msg) == 1
 }
 
 #' Set up application ID and app secret.
@@ -102,24 +124,35 @@ secret_exists <- function(secret = c(TOKEN, APP_ID, APP_SECRET)) {
 #' @importFrom utils menu
 adlib_setup <- function() {
   get_id <- TRUE
-  if (secret_exists(APP_ID)) {
-    get_id <- menu(c("y", "n"), title = "Application ID already set. Overwrite? (y/n)") == 1
+  get_secret <- TRUE
+  backend_is_env <- keyring_backend_is_env()
+  storage_space <- "in your system's secure credential store."
+
+  if (backend_is_env) {
+    if (!yesno("Your system does not have a secure credential store. Secrets will be stored as environment variables. Delete them at any time with adlib_clear_setup(). Continue?")) {
+      stop("User aborted setup.")
+    }
+    storage_space <- "as environment variables"
   }
+
+  if (secret_exists(APP_ID)) {
+    get_id <- yesno("App ID already set. Overwrite?")
+  }
+
   if (get_id) {
-    message("Visit https://developers.facebook.com/ and navigate to your App's basic settings
-to find your Application ID and App Secret.
-These will be securely stored in your computer's credential store.")
+    message(glue::glue("Visit https://developers.facebook.com/ and navigate to your App's basic settings
+to find your Application ID and App Secret. These will be stored {storage_space}."))
     readline("Press <Enter> ")
-    app_id <- getPass::getPass("Enter your Application ID", noblank = T, forcemask = F)
+    app_id <- get_pass("Enter your Application ID")
     secret_set(APP_ID, app_id)
   }
 
-  get_secret <- TRUE
   if (secret_exists(APP_SECRET)) {
-    get_secret <- menu(c("y", "n"), title = "App secret already set. Overwrite? (y/n)") == 1
+    get_secret <- yesno("App secret already set. Overwrite?")
   }
+
   if (get_secret) {
-    app_secret <- getPass::getPass("Enter your App secret", noblank = T, forcemask = F)
+    app_secret <- get_pass("Enter your App secret")
     secret_set(APP_SECRET, app_secret)
   }
   message("Application ID and App secret set. Run adlib_set_longterm_token() to save a long term access token.")