+++ NEWS.md
@@ -1,3 +1,8 @@
+# Radlibrary 0.2.4
+* Fixed getPass bug in Rstudio server
+* Only warn once about environment variables on Linux
+* fix typos in docs
+
 # Radlibrary 0.2.3
 * improve docs
 * pass rhub windows build check